requirements.txt
Flask 
mysql-connector-python
fastapi
pydantic